refac file struc & add to ci tooling cfr Sam's feedback
