+++ .github/workflows/ci.yml
@@ -9,10 +9,15 @@ jobs:
       - uses: actions/setup-go@v2
         with:
           go-version: "^1.16"
+      - name: run gofmt
+        run: test -z $(gofmt -l .)
+      - name: run go vet
+        run: go vet ./...
+      - uses: golangci/golangci-lint-action@v2
       - name: golangci-lint
-        uses: golangci/golangci-lint-action@v2
-      - run: golangci-lint run ./...
-      - run: go test ./... -coverprofile=cover.out -covermode=atomic
+        run: golangci-lint run ./...
+      - name: run tests
+        run: go test ./... -coverprofile=cover.out -covermode=atomic
       - uses: codecov/codecov-action@v2
         with:
           files: ./cover.out
